Extract response utils
We have some methods in the authenticator package that handle HTTP
responses. We can benefit from extracting them into utils with code
clarity.
